backend/requirements.txt
flask
flask-cors
gunicorn
joblib
scikit-learn
numpy
scipy
nltk
textblob
requests
pdfplumber
pytesseract
pdf2image
reportlab
newspaper3k
lxml
lxml_html_clean
beautifulsoup4
ddgs
readability-lxml
